+++ R/parsnip.R
@@ -55,11 +55,16 @@ orbital.model_fit <- function(x, ..., prefix = ".pred", type = NULL) {
 	}
 
 	attr(res, "pred_names") <- names
-	if (inherits(x, "_xgb.Booster")) {
-		na_fields <- which(is.na(names(res)))
-		tmp_names <- names(res)
-		tmp_names[na_fields] <- paste0(prefix, "_", x$lvl)
-		names(res) <- tmp_names
+	if (
+		inherits(x, "_xgb.Booster") &&
+			isTRUE(x$fit$params$objective == "multi:softprob")
+	) {
+		if (anyNA(names(res))) {
+			na_fields <- which(is.na(names(res)))
+			tmp_names <- names(res)
+			tmp_names[na_fields] <- paste0(prefix, "_", x$lvl)
+			names(res) <- tmp_names
+		}
 	} else {
 		res <- stats::setNames(res, names)
 	}