+++ NEWS.md
@@ -1,4 +1,4 @@
-# orbital (development version)
+# orbital 0.3.0
 
 * `orbital()` has gained `type` argument to change prediction type. (#66)